ecs/DQ-agent/requirements.txt
strands-agents==1.0.0
fastapi==0.115.12
pydantic==2.11.5
mcp==1.11.0
uvicorn==0.34.3
pyjwt==2.10.1
cryptography==45.0.4
requests==2.32.3
